出品者以外のユーザーを弾く

+++ app/controllers/items_controller.rb
@@ -38,6 +38,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
+
+    if item.user != current_user
+      redirect_to root_path
+      return
+    end
+
     item.destroy
     redirect_to root_path
   end
